Minor change to type for eigenvector

+++ 3DRenderer/forms/mesh.h
@@ -11,7 +11,7 @@
  */
 struct Vertex {
     Eigen::Vector3f position;
-    Eigen::Vector3d color;
+    Eigen::Vector3i color;
     // Eigen::Vector3f normal;
     // Eigen::Vector2f TexCoords;
 };
